backend/requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile --output-file=base.txt base.in
#
advocate==1.0.0
    # via -r base.in
aioredis==1.3.1
    # via channels-redis
amqp==5.1.1
    # via kombu
antlr4-python3-runtime==4.9.3
    # via -r base.in
anyio==3.6.1
    # via watchgod
asgiref==3.4.1
    # via
    #   -r base.in
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   uvicorn
async-timeout==4.0.2
    # via aioredis
attrs==22.1.0
    # via
    #   automat
    #   jsonschema
    #   service-identity
    #   twisted
autobahn==22.7.1
    # via daphne
automat==20.2.0
    # via twisted
billiard==3.6.4.0
    # via celery
boto3==1.20.38
    # via -r base.in
botocore==1.23.54
    # via
    #   boto3
    #   s3transfer
celery[redis]==5.2.3
    # via
    #   -r base.in
    #   celery-redbeat
    #   django-celery-beat
    #   django-celery-email
celery-redbeat==2.0.0
    # via -r base.in
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
channels==3.0.4
    # via
    #   -r base.in
    #   channels-redis
channels-redis==3.3.0
    # via -r base.in
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
constantly==15.1.0
    # via twisted
cryptography==38.0.1
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==3.0.2
    # via channels
deprecated==1.2.13
    # via redis
dj-database-url==0.5.0
    # via -r base.in
django==3.2.13
    # via
    #   -r base.in
    #   channels
    #   django-appconf
    #   django-celery-beat
    #   django-celery-email
    #   django-cors-headers
    #   django-health-check
    #   django-redis
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-appconf==1.0.5
    # via django-celery-email
django-celery-beat==2.2.1
    # via -r base.in
django-celery-email==3.0.0
    # via -r base.in
django-cors-headers==3.8.0
    # via -r base.in
django-health-check==3.16.5
    # via -r base.in
django-redis==5.2.0
    # via -r base.in
django-storages==1.12.3
    # via -r base.in
django-timezone-field==4.2.3
    # via django-celery-beat
djangorestframework==3.13.1
    # via
    #   -r base.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.2.0
    # via -r base.in
drf-spectacular==0.21.2
    # via -r base.in
faker==8.11.0
    # via -r base.in
gunicorn==20.1.0
    # via -r base.in
h11==0.14.0
    # via uvicorn
hiredis==2.0.0
    # via aioredis
httptools==0.2.0
    # via uvicorn
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.4
    # via
    #   anyio
    #   hyperlink
    #   requests
    #   twisted
incremental==21.3.0
    # via twisted
inflection==0.5.1
    # via drf-spectacular
itsdangerous==2.0.1
    # via -r base.in
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==4.16.0
    # via drf-spectacular
kombu==5.2.4
    # via celery
msgpack==1.0.4
    # via channels-redis
ndg-httpsclient==0.5.1
    # via advocate
netifaces==0.11.0
    # via advocate
packaging==21.3
    # via redis
pillow==9.0.0
    # via -r base.in
prompt-toolkit==3.0.31
    # via click-repl
psutil==5.9.0
    # via -r base.in
psycopg2==2.9.1
    # via -r base.in
pyasn1==0.4.8
    # via
    #   advocate
    #   ndg-httpsclient
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.5.0
    # via drf-jwt
pyopenssl==22.1.0
    # via
    #   advocate
    #   ndg-httpsclient
    #   twisted
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-crontab==2.6.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   botocore
    #   celery-redbeat
    #   faker
    #   python-crontab
python-dotenv==0.21.0
    # via uvicorn
pytz==2022.4
    # via
    #   celery
    #   django
    #   django-timezone-field
    #   djangorestframework
pyyaml==6.0
    # via
    #   drf-spectacular
    #   uvicorn
redis==4.1.4
    # via
    #   -r base.in
    #   celery
    #   celery-redbeat
    #   django-redis
regex==2021.8.3
    # via -r base.in
requests==2.26.0
    # via
    #   -r base.in
    #   advocate
s3transfer==0.5.2
    # via boto3
service-identity==21.1.0
    # via
    #   -r base.in
    #   twisted
six==1.16.0
    # via
    #   advocate
    #   automat
    #   click-repl
    #   python-dateutil
    #   service-identity
sniffio==1.3.0
    # via anyio
sqlparse==0.4.3
    # via django
tenacity==8.1.0
    # via celery-redbeat
text-unidecode==1.3
    # via faker
tqdm==4.62.3
    # via -r base.in
twisted[tls]==22.1.0
    # via
    #   -r base.in
    #   daphne
txaio==22.2.1
    # via autobahn
types-cryptography==3.3.23
    # via pyjwt
typing-extensions==4.4.0
    # via twisted
unicodecsv==0.14.1
    # via -r base.in
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.12
    # via
    #   advocate
    #   botocore
    #   requests
uvicorn[standard]==0.15.0
    # via -r base.in
uvloop==0.17.0
    # via uvicorn
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchgod==0.8.2
    # via uvicorn
wcwidth==0.2.5
    # via prompt-toolkit
websockets==10.3
    # via
    #   -r base.in
    #   uvicorn
wrapt==1.14.1
    # via deprecated
zipp==3.5.0
    # via -r base.in
zope-interface==5.4.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
